Fix type references following refactor
This type moved to a different file. I debated creating an `index.ts`
file to simplify imports, or re-exporting the type from the old file so
as not to change imports, but I opted to do this the simplest and most
straightforward way.
