+++ x-pack/solutions/security/plugins/security_solution/server/lib/detection_engine/rule_management/utils/utils.ts
@@ -29,7 +29,7 @@ import { createBulkErrorObject } from '../../routes/utils';
 import type { InvestigationFieldsCombined, RuleAlertType, RuleParams } from '../../rule_schema';
 import { hasValidRuleType } from '../../rule_schema';
 import { internalRuleToAPIResponse } from '../logic/detection_rules_client/converters/internal_rule_to_api_response';
-import type { BulkActionError } from '../api/rules/bulk_actions/bulk_actions_response';
+import type { BulkActionError } from '../api/rules/bulk_actions/utils';
 
 type PromiseFromStreams = RuleToImport | Error;
 const MAX_CONCURRENT_SEARCHES = 10;
